Roles list

+++ backend/src/main/java/com/techelevator/model/ShiftRoleDAO.java
@@ -7,9 +7,9 @@ import com.techelevator.pojo.ShiftRole;
 public interface ShiftRoleDAO {
 	
 	public boolean saveShiftRole(long companyId, String title);
-	public List<ShiftRole> getAllShiftRoles(int companyId);
+	public List<ShiftRole> getAllShiftRoles(long companyId);
 	public ShiftRole getShiftRoleById(int id);
-	public boolean clearAllShiftRoles(int companyId);
+	public boolean clearAllShiftRoles(long companyId);
 	public int getNextShiftRoleId();
 
 }
